reformat file structure of optim
