+++ pypose/optim/__init__.py
@@ -1,2 +1,3 @@
 from . import functional
-from .optimizer import *
+from .optimizer import GaussNewton, LevenbergMarquardt
+from .optimizer import GaussNewton as GN, LevenbergMarquardt as LM
