--- pypose/optim/optimizer/__init__.py
@@ -1,4 +1,0 @@
-from .gauss_newton import GaussNewton
-from .gauss_newton import GaussNewton as GN
-from .levenberg_marquardt import LevenbergMarquardt
-from .levenberg_marquardt import LevenbergMarquardt as LM
